Downloads/docker-infra-task/docker-infra-task/backend/requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2

#My added dependencies
mysqlclient==2.2.4
python-dotenv==1.0.1
